added load_wrapper and try_resolve_uri benchmarks
