+++ packages/benchmarks/Cargo.toml
@@ -1,27 +1,28 @@
 [package]
 name = "polywrap_benchmarks"
 description = "Polywrap Client benchmarks in rust"
-
-version.workspace = true
-edition.workspace = true
-rust-version.workspace = true
-license.workspace = true
+version = "0.1.6-beta.4"
+license = "MIT"
+edition = "2021"
+rust-version = "1.67"
+repository = "https://github.com/polywrap/rust-client"
 
 [lib]
+crate-type = ["rlib"]
 test = false
 
 [dependencies]
-polywrap_client.workspace = true
-polywrap_wasm.workspace = true
-polywrap_client_builder.workspace = true
-polywrap_core.workspace = true
-wrap_manifest_schemas.workspace = true
-polywrap_resolvers.workspace = true
-polywrap_msgpack.workspace = true
-polywrap_tests_utils.workspace = true
-polywrap_plugin.workspace = true
-polywrap_plugin_implementor.workspace = true
-polywrap_client_default_config.workspace = true
+polywrap_client = { version = "0.1.6-beta.4" }
+polywrap_wasm = { version = "0.1.6-beta.4" }
+polywrap_client_builder = { version = "0.1.6-beta.4" }
+polywrap_core = { version = "0.1.6-beta.4" }
+wrap_manifest_schemas = { version = "0.1.6-beta.4" }
+polywrap_resolvers = { version = "0.1.6-beta.4" }
+polywrap_msgpack = { version = "0.1.6-beta.4" }
+polywrap_tests_utils = { version = "0.1.6-beta.4" }
+polywrap_plugin = { version = "0.1.6-beta.4" }
+polywrap_plugin_implementor = { version = "0.1.6-beta.4" }
+polywrap_client_default_config = { version = "0.1.6-beta.4" }
 
 [dev-dependencies]
 criterion = { version = "0.4", features = ["html_reports"] }
